# openapi.yaml
openapi: "3.0.3"
info:
  title: spacey-access
  version: "0.1.0"
  description: >
    Access service for spacey. Manages access codes and room check-in/check-out.
    Purchase owns the paid check; Access never reads the bookings table.

paths:
  /health:
    get:
      summary: Service and database health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  revision:
                    type: string
        "503":
          description: Database unreachable

  /openapi.yaml:
    get:
      summary: This contract
      responses:
        "200":
          description: YAML document

  # --- Purchase -> Access -------------------------------------------------
  # Responses are DUMMY for now (fixed values); 403/404/409 come with the real logic.

  /bookings/{booking_id}/access:
    post:
      summary: Create the access record for a paid booking (Purchase)
      description: Purchase checks payment before calling this.
      security:
        - ServiceToken: []
      parameters:
        - $ref: "#/components/parameters/BookingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [start_time, end_time]
              properties:
                start_time: { type: string, format: date-time }
                end_time: { type: string, format: date-time }
      responses:
        "201":
          description: Access created
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id: { type: integer }
                  access_code: { type: string, example: "00000000" }
                  status: { type: string, example: available }
                  expires_at: { type: string, format: date-time }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /bookings/{booking_id}/access/remove:
    post:
      summary: Remove access, the booking was cancelled (Purchase)
      description: Soft delete. The record is kept with status removed.
      security:
        - ServiceToken: []
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Access removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id: { type: integer }
                  status: { type: string, example: removed }
                  removed_at: { type: string, format: date-time }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /bookings/{booking_id}/access/expire:
    post:
      summary: Expire access, the booking has ended by Purchase's clock (Purchase)
      security:
        - ServiceToken: []
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Access expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id: { type: integer }
                  status: { type: string, example: expired }
        "401":
          $ref: "#/components/responses/Unauthorized"

  # --- Frontend -> Access -------------------------------------------------

  /bookings/{booking_id}/check-in:
    post:
      summary: Check in with the access code (Frontend)
      parameters:
        - $ref: "#/components/parameters/BookingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_code]
              properties:
                access_code: { type: string, example: "00000000" }
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id: { type: integer }
                  status: { type: string, example: used }
                  checked_in_at: { type: string, format: date-time }
        "400":
          $ref: "#/components/responses/BadRequest"

  /bookings/{booking_id}/check-out:
    post:
      summary: Check out (Frontend). Access becomes available again until it expires
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id: { type: integer }
                  status: { type: string, example: available }
                  checked_out_at: { type: string, format: date-time }

# Access -> Purchase (not part of this API): when access expires by Access's clock,
# Access calls POST {PURCHASE_URL}/bookings/{booking_id}/access-expired
# (src/purchase_client.py; dummy, nothing is sent until Purchase has that endpoint).

components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: >
        Shared secret between Purchase and Access (SERVICE_TOKEN), sent as
        `Authorization: Bearer <token>`. Required on create, remove and expire.
        The value lives in the platform and CI secrets, never in Git.

  parameters:
    BookingId:
      name: booking_id
      in: path
      required: true
      schema: { type: integer }

  schemas:
    Error:
      type: object
      properties:
        error: { type: string }

  responses:
    Unauthorized:
      description: The service token is missing or wrong
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: A required field is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"